add dynamic loading for TestBehaviour
