--- supplementary/alica_ros1/alica_tests/src/test/Expr/include/alica_tests/TestBehaviour.h
@@ -1,30 +1,0 @@
-#pragma once
-
-#include <alica_tests/DomainBehaviour.h>
-/*PROTECTED REGION ID(inc55178365253414982) ENABLED START*/
-// Add additional includes here
-/*PROTECTED REGION END*/
-
-namespace alica
-{
-class TestBehaviour : public DomainBehaviour
-{
-public:
-    TestBehaviour(BehaviourContext& context);
-    virtual ~TestBehaviour();
-    virtual void run();
-    /*PROTECTED REGION ID(pub55178365253414982) ENABLED START*/
-    // Add additional protected methods here
-    /*PROTECTED REGION END*/
-protected:
-    virtual void initialiseParameters();
-    /*PROTECTED REGION ID(pro55178365253414982) ENABLED START*/
-    // Add additional protected methods here
-    virtual void onTermination();
-    /*PROTECTED REGION END*/
-private:
-    /*PROTECTED REGION ID(prv55178365253414982) ENABLED START*/
-    // Add additional private methods here
-    /*PROTECTED REGION END*/
-};
-} /* namespace alica */
