--- supplementary/alica_ros1/alica_tests/src/test/Expr/include/alica_tests/TestBehaviour55178365253414982.h
@@ -1,14 +1,0 @@
-
-
-#include <alica_tests/DomainCondition.h>
-/*PROTECTED REGION ID(incBC55178365253414982) ENABLED START*/
-// Add additional includes here
-/*PROTECTED REGION END*/
-
-namespace alica
-{
-/*PROTECTED REGION ID(meth55178365253414982) ENABLED START*/
-// Add additional includes here
-/*PROTECTED REGION END*/
-
-} /* namespace alica */
