--- supplementary/alica_ros1/alica_tests/src/test/Expr/src/TestBehaviour55178365253414982.cpp
@@ -1,12 +1,0 @@
-#include <alica_tests/TestBehaviour55178365253414982.h>
-#include <memory>
-
-/*PROTECTED REGION ID(inccppBC55178365253414982) ENABLED START*/
-// Add additional includes here
-/*PROTECTED REGION END*/
-
-namespace alica
-{
-// Behaviour:TestBehaviour
-
-} /* namespace alica */
